Allow to disable depth camera + move parameter node to tango_common
