+++ TangoRosStreamer/app/src/main/java/eu/intermodalics/tango_ros_streamer/activities/SettingsActivity.java
@@ -119,6 +119,7 @@ public class SettingsActivity extends AppCompatPreferenceActivity implements
         if (key == getString(R.string.pref_master_is_local_key) ||
                 key == getString(R.string.pref_master_uri_key) ||
                 key == getString(R.string.pref_create_new_map_key) ||
+                key == getString(R.string.pref_enable_depth_key) ||
                 key == getString(R.string.pref_localization_mode_key) ||
                 key == getString(R.string.pref_localization_map_uuid_key)) {
             boolean previouslyStarted = mSharedPref.getBoolean(getString(R.string.pref_previously_started_key), false);
@@ -133,8 +134,9 @@ public class SettingsActivity extends AppCompatPreferenceActivity implements
                 }
                 // These changes require to restart Tango only.
                 if (key == getString(R.string.pref_create_new_map_key) ||
-                        key == getString(R.string.pref_localization_mode_key) ||
-                        key == getString(R.string.pref_localization_map_uuid_key)) {
+                    key == getString(R.string.pref_enable_depth_key) ||
+                    key == getString(R.string.pref_localization_mode_key) ||
+                    key == getString(R.string.pref_localization_map_uuid_key)) {
                     Snackbar snackbar = Snackbar.make(mSettingsPreferenceFragment.getView(), getString(R.string.snackbar_text_restart_tango), Snackbar.LENGTH_INDEFINITE);
                     snackbar.setAction(getString(R.string.snackbar_action_text_restart_tango), new View.OnClickListener() {
                         @Override
