+++ TangoRosStreamer/app/src/main/res/values/strings.xml
@@ -74,6 +74,10 @@
     <string name="pref_create_new_map_key">create_new_map</string>
     <string name="pref_create_new_map_summary">Enable to create a new localization map. This map can be used for localization later.</string>
 
+    <string name="pref_enable_depth">Enable depth camera</string>
+    <string name="pref_enable_depth_key">enable_depth_camera</string>
+    <string name="pref_enable_depth_summary">Enable depth camera when connecting to Tango.</string>
+
     <string name="pref_localization_mode_key">localization_mode</string>
     <string name="pref_localization_map_uuid_key">localization_map_uuid</string>
